Change name of lambda variable
Changed the name of the lambda variable from "a_proc" to "a_lambda" in the "Procs vs. lambdas"-"Arguments" section for clarity.
I'm aware that a lambda is in a fact also a proc, but it goes against the wording in the rest of the lessson (see the very next section, so I assume it's an oversight.